feat: :truck: restructure src/core directory
